infinite glue is now immune from acetone

+++ code/modules/chemistry/Reagents-Misc.dm
@@ -1022,7 +1022,8 @@ datum
 				var/atom/Aloc = isturf(A) ? A : A.loc
 				for(var/atom/movable/AM in Aloc)
 					var/datum/component/glued/glued_comp = AM.GetComponent(/datum/component/glued)
-					if(glued_comp?.glued_to == A)
+					// possible idea for a future change: instead of direct deletion just decrease glue_removal_time and only delete if <= 0
+					if(glued_comp?.glued_to == A && !isnull(glued_comp.glue_removal_time))
 						qdel(glued_comp)
 
 			reaction_mob(var/mob/M, var/method=TOUCH, var/volume, var/paramslist = 0, var/raw_volume)
